Fix merge #JC19-427
