+++ apps/angular/src/app/features/auth/login/login.component.ts
@@ -5,7 +5,7 @@ import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms'
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from '@angular/core';
 
 import { Login } from '@js-camp/core/models/login';
-import { AppError } from '@js-camp/core/models/httpError';
+import { AppError } from '@js-camp/core/models/app-error';
 import { FormError } from '@js-camp/core/models/form-error';
 
 import { UrlService } from '../../../../core/services/url.service';
