+++ libs/core/models/http-error.ts
@@ -1,0 +1,13 @@
+/** Error coming from the server. */
+export class HttpError extends Error {
+
+  /**
+   * Error message.
+   */
+  public override readonly message: string;
+
+  public constructor(message: string) {
+    super(message);
+    this.message = message;
+  }
+}
